% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize strings}
\usage{
capitalize(x, force_tolower = FALSE)
}
\arguments{
\item{x}{Character vector to capitalize}

\item{force_tolower}{Logical, whether to convert to lowercase first}
}
\value{
Character vector with capitalized strings
}
\description{
Capitalizes the first letter of each word in a string
}
