% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-imputation.R
\name{choose_k}
\alias{choose_k}
\title{Helper function to choose optimal k for ALRA}
\usage{
choose_k(d, q = 0.01)
}
\arguments{
\item{d}{Singular values}

\item{q}{Quantile threshold}
}
\value{
Optimal k value
}
\description{
Helper function to choose optimal k for ALRA
}
