% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-cache.R
\name{clear_SCP_cache}
\alias{clear_SCP_cache}
\title{Clear SCP cache}
\usage{
clear_SCP_cache(older_than = NULL)
}
\arguments{
\item{older_than}{Remove only files older than this many days (NULL = all files)}
}
\description{
Remove cached external data files
}
