% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2hex}
\alias{col2hex}
\title{Convert color to hex format}
\usage{
col2hex(colors)
}
\arguments{
\item{colors}{Character vector of color names or specifications}
}
\value{
Character vector of hex color codes
}
\description{
Replacement for gplots::col2hex using base R functions
}
\keyword{internal}
