% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration-utils.R
\name{.check_scaled_v5}
\alias{.check_scaled_v5}
\title{Check if features are scaled without joining layers (V5 optimized)}
\usage{
.check_scaled_v5(srt, features, assay)
}
\arguments{
\item{srt}{Seurat object}

\item{features}{Features to check}

\item{assay}{Assay name}
}
\value{
Logical, TRUE if all features are scaled
}
\description{
Check if features are scaled without joining layers (V5 optimized)
}
