% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration-utils.R
\name{.convert_layers_to_list}
\alias{.convert_layers_to_list}
\title{Convert V5 layers back to list of objects}
\usage{
.convert_layers_to_list(srt, batch, assay)
}
\arguments{
\item{srt}{Seurat v5 object with layers}

\item{batch}{Character, batch variable name}

\item{assay}{Character, assay name}
}
\value{
List of Seurat objects
}
\description{
Convert V5 layers back to list of objects
}
