% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration-utils.R
\name{.convert_list_to_layers}
\alias{.convert_list_to_layers}
\title{Convert list of Seurat objects to V5 layers}
\usage{
.convert_list_to_layers(srtList, batch, assay)
}
\arguments{
\item{srtList}{List of Seurat objects}

\item{batch}{Character, batch variable name}

\item{assay}{Character, assay name}
}
\value{
Seurat v5 object with layers split by batch
}
\description{
Convert list of Seurat objects to V5 layers
}
