% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration-utils.R
\name{.post_process_integration}
\alias{.post_process_integration}
\title{Post-process integration results based on workflow}
\usage{
.post_process_integration(srtIntegrated, workflow, assay)
}
\arguments{
\item{srtIntegrated}{Integrated Seurat object}

\item{workflow}{Character, "layers" or "list"}

\item{assay}{Character, assay name}
}
\value{
Processed Seurat object
}
\description{
Post-process integration results based on workflow
}
