% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration-utils.R
\name{.prepare_integration_inputs}
\alias{.prepare_integration_inputs}
\title{Prepare integration inputs based on workflow}
\usage{
.prepare_integration_inputs(srtList, srtMerge, batch, assay, workflow)
}
\arguments{
\item{srtList}{List of Seurat objects or NULL}

\item{srtMerge}{Seurat object or NULL}

\item{batch}{Character, batch variable name}

\item{assay}{Character, assay name}

\item{workflow}{Character, "layers" or "list"}
}
\value{
List with prepared inputs
}
\description{
Prepare integration inputs based on workflow
}
