% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_pyproject_toml}
\alias{find_pyproject_toml}
\title{Find pyproject.toml file}
\usage{
find_pyproject_toml()
}
\value{
Character path to pyproject.toml or NULL if not found
}
\description{
Find pyproject.toml file
}
\keyword{internal}
