% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-analysis.R
\name{FindExpressedMarkers}
\alias{FindExpressedMarkers}
\title{Find expressed marker features}
\usage{
FindExpressedMarkers(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  features = NULL,
  assay = NULL,
  slot = "data",
  min.expression = 0,
  test.use = "wilcox",
  logfc.threshold = 0.25,
  base = 2,
  pseudocount.use = 1,
  mean.fxn = NULL,
  fc.name = NULL,
  min.pct = 0.1,
  min.diff.pct = -Inf,
  max.cells.per.ident = Inf,
  latent.vars = NULL,
  only.pos = FALSE,
  min.cells.group = 3,
  min.cells.feature = 3,
  norm.method = "LogNormalize",
  verbose = TRUE,
  ...
)
}
\description{
Find features that are expressed in specified groups of cells
}
\examples{
markers <- FindExpressedMarkers(pancreas_sub, cells.1 = WhichCells(pancreas_sub, expression = Phase == "G2M"))
head(markers)
FeatureStatPlot(pancreas_sub, rownames(markers)[1], group.by = "Phase", add_point = TRUE)
}
