% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_feature_metadata}
\alias{get_feature_metadata}
\title{Get feature metadata from Seurat object in a version-agnostic way}
\usage{
get_feature_metadata(srt, assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{assay}{Name of the assay. If NULL, uses the default assay.}
}
\value{
A data frame of feature metadata
}
\description{
This function extracts feature metadata from a Seurat object, handling both V4 and V5 formats.
}
