% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_scp_pkg_dir}
\alias{get_scp_pkg_dir}
\title{Get SCP package directory}
\usage{
get_scp_pkg_dir()
}
\value{
Character, path to SCP package directory
}
\description{
Get SCP package directory
}
\keyword{internal}
