% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-conversion.R
\name{get_spatial_coords}
\alias{get_spatial_coords}
\title{Get spatial coordinates from Seurat object}
\usage{
get_spatial_coords(srt, image = NULL)
}
\arguments{
\item{srt}{Seurat object with spatial data}

\item{image}{Image name to extract coordinates from (default: first image)}
}
\value{
Data frame with spatial_x and spatial_y columns, rownames are cell barcodes
}
\description{
Extracts spatial coordinates from either the Images slot or metadata
}
\examples{
\dontrun{
coords <- get_spatial_coords(srt)
head(coords)
}

}
