% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_features}
\alias{get_var_features}
\title{Get variable features from Seurat object in a version-agnostic way}
\usage{
get_var_features(srt, assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{assay}{Name of the assay. If NULL, uses the default assay.}
}
\value{
A character vector of variable feature names
}
\description{
This function extracts variable features from a Seurat object, handling both V4 and V5 formats.
}
