% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_uv}
\alias{install_uv}
\title{Install UV if not present}
\usage{
install_uv(force = FALSE)
}
\arguments{
\item{force}{Force reinstallation even if UV exists}
}
\description{
Install UV if not present
}
