% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependencies.R
\name{invoke}
\alias{invoke}
\title{Helper function to invoke methods dynamically}
\usage{
invoke(.fn, .args = list(), ...)
}
\arguments{
\item{.fn}{Function to invoke (can be function object, name, or namespace::function)}

\item{.args}{List of arguments to pass to the function}

\item{...}{Additional arguments (will be added to .args)}
}
\value{
Result of the invoked function
}
\description{
This function provides a clean interface for dynamically calling functions
with a list of arguments, similar to do.call but with a more flexible syntax.
}
