% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-workflow.R
\name{is_seurat_v5}
\alias{is_seurat_v5}
\title{Check if a Seurat object is version 5}
\usage{
is_seurat_v5(srt)
}
\arguments{
\item{srt}{An object of class 'Seurat'}
}
\value{
Logical indicating whether the object is Seurat v5
}
\description{
Check if a Seurat object is version 5
}
