% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-conversion.R
\name{is_spatial_object}
\alias{is_spatial_object}
\title{Check if Seurat object has spatial data}
\usage{
is_spatial_object(srt)
}
\arguments{
\item{srt}{Seurat object}
}
\value{
Logical indicating presence of spatial data
}
\description{
Simple boolean check for spatial data presence
}
\examples{
\dontrun{
if (is_spatial_object(srt)) {
  message("This is a spatial dataset")
}
}

}
