% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsSeurat5}
\alias{IsSeurat5}
\title{Check if object is Seurat V5}
\usage{
IsSeurat5(srt)
}
\arguments{
\item{srt}{A Seurat object}
}
\value{
Logical indicating if the object is Seurat V5
}
\description{
Check if object is Seurat V5
}
