% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-imputation.R
\name{matrix_power}
\alias{matrix_power}
\title{Helper function for matrix power}
\usage{
matrix_power(mat, power)
}
\arguments{
\item{mat}{Input matrix}

\item{power}{Power to raise matrix to}
}
\value{
Matrix raised to specified power
}
\description{
Helper function for matrix power
}
