% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-plot.R
\name{print.upset_plot_scp}
\alias{print.upset_plot_scp}
\title{Print method for upset plots}
\usage{
\method{print}{upset_plot_scp}(x, ...)
}
\arguments{
\item{x}{An upset_plot_scp object}

\item{...}{Additional arguments passed to print.ggplot}
}
\description{
This print method suppresses harmless warnings from ggupset package that occur
when rendering upset plots (warnings about geom_line groups and deprecated size aesthetic).
}
\keyword{internal}
