% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-squidpy.R
\name{RunCentralityScores}
\alias{RunCentralityScores}
\title{Compute centrality scores}
\usage{
RunCentralityScores(
  srt,
  cluster_key = "seurat_clusters",
  assay = NULL,
  verbose = TRUE
)
}
\arguments{
\item{srt}{Seurat object with spatial graph}

\item{cluster_key}{Column name in metadata containing cluster labels}

\item{assay}{Assay to use}

\item{verbose}{Print progress messages}
}
\value{
Seurat object with centrality scores in misc slot
}
\description{
Computes centrality scores for clusters in the spatial graph.
}
\examples{
\dontrun{
srt <- RunSpatialGraph(srt, method = "knn")
srt <- RunCentralityScores(srt, cluster_key = "seurat_clusters")
}
}
