% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-imputation.R
\name{RunImputation}
\alias{RunImputation}
\title{Run imputation methods on Seurat object}
\usage{
RunImputation(
  srt,
  assay = NULL,
  slot = "data",
  method = c("alra", "magic", "knn_smooth"),
  new_assay = NULL,
  features = NULL,
  verbose = TRUE
)
}
\arguments{
\item{srt}{An object of class Seurat.}

\item{assay}{Name of assay to use for imputation.}

\item{slot}{Name of slot to use for imputation (default: "data").}

\item{method}{Imputation method to use. One of "alra", "magic", "knn_smooth".}

\item{new_assay}{Name of the new assay to store imputed data (default: paste0(method, "_imputed")).}

\item{features}{Features to use for imputation. If NULL, uses variable features.}

\item{verbose}{Print progress messages.}
}
\value{
A Seurat object with imputed values stored in a new assay.
}
\description{
This function performs imputation on a Seurat object using various methods including ALRA, MAGIC, etc.
The imputed values are stored as a new assay in the Seurat object.
}
