% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-imputation.R
\name{RunMAGIC}
\alias{RunMAGIC}
\title{Run MAGIC imputation}
\usage{
RunMAGIC(data_mat, t = 3, k = 10, verbose = TRUE)
}
\arguments{
\item{data_mat}{Expression matrix with genes as rows and cells as columns}

\item{t}{Time parameter for diffusion (default: 3)}

\item{k}{Number of nearest neighbors (default: 10)}

\item{verbose}{Print progress messages}
}
\value{
Imputed expression matrix
}
\description{
This function performs MAGIC (Markov Affinity-based Graph Imputation of Cells)
imputation on a gene expression matrix.
}
