% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-squidpy.R
\name{RunSpatialAutocorr}
\alias{RunSpatialAutocorr}
\title{Compute spatial autocorrelation}
\usage{
RunSpatialAutocorr(
  srt,
  mode = c("moran", "geary"),
  genes = NULL,
  n_perms = 1000,
  n_jobs = 1,
  assay = NULL,
  verbose = TRUE
)
}
\arguments{
\item{srt}{Seurat object with spatial graph}

\item{mode}{Autocorrelation metric: "moran" (Moran's I) or "geary" (Geary's C)}

\item{genes}{List of genes to test (default: all variable features)}

\item{n_perms}{Number of permutations (default: 1000)}

\item{n_jobs}{Number of parallel jobs (default: 1)}

\item{assay}{Assay to use}

\item{verbose}{Print progress messages}
}
\value{
Seurat object with autocorrelation results in misc slot
}
\description{
Computes spatial autocorrelation statistics (Moran's I or Geary's C) for genes.
}
\examples{
\dontrun{
srt <- RunSpatialGraph(srt, method = "knn")
srt <- RunSpatialAutocorr(srt, mode = "moran")
}
}
