% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_feature_metadata}
\alias{set_feature_metadata}
\title{Set feature metadata in Seurat object in a version-agnostic way}
\usage{
set_feature_metadata(srt, metadata, assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{metadata}{A data frame of feature metadata}

\item{assay}{Name of the assay. If NULL, uses the default assay.}
}
\value{
The modified Seurat object
}
\description{
This function sets feature metadata in a Seurat object, handling both V4 and V5 formats.
}
