% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_seurat_data}
\alias{set_seurat_data}
\title{Set data in Seurat object in a version-agnostic way}
\usage{
set_seurat_data(srt, data, layer = "data", assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{data}{The data to set}

\item{layer}{The layer to set (e.g., "counts", "data", "scale.data")}

\item{assay}{Name of the assay. If NULL, uses the default assay.}
}
\value{
The modified Seurat object
}
\description{
This function sets data in a Seurat object, handling both V4 and V5 formats.
}
