% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-plot.R
\name{SpatialCoOccurrencePlot}
\alias{SpatialCoOccurrencePlot}
\title{Plot spatial co-occurrence}
\usage{
SpatialCoOccurrencePlot(srt, cluster_pairs = NULL, ...)
}
\arguments{
\item{srt}{Seurat object with co-occurrence results}

\item{cluster_pairs}{Specific cluster pairs to plot (optional)}

\item{...}{Additional parameters}
}
\value{
ggplot object
}
\description{
Visualizes spatial co-occurrence probability between clusters
}
\examples{
\dontrun{
srt <- RunSpatialCoOccurrence(srt)
SpatialCoOccurrencePlot(srt)
}
}
