% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCP-spatial-plot.R
\name{SpatialFeaturePlot}
\alias{SpatialFeaturePlot}
\title{Spatial feature plot}
\usage{
SpatialFeaturePlot(
  srt,
  features,
  image = NULL,
  cols = c("lightgrey", "red"),
  alpha = 1,
  pt.size = 1.6,
  ncol = NULL,
  combine = TRUE,
  ...
)
}
\arguments{
\item{srt}{Seurat object with spatial data}

\item{features}{Features to plot (genes or metadata columns)}

\item{image}{Image name to use (default: first image)}

\item{cols}{Color scale for features}

\item{alpha}{Transparency (0-1)}

\item{pt.size}{Point size}

\item{ncol}{Number of columns for multi-feature plots}

\item{combine}{Combine plots into single figure}

\item{...}{Additional parameters passed to Seurat::SpatialFeaturePlot}
}
\value{
ggplot object or list of ggplot objects
}
\description{
Visualizes gene expression or other features on spatial coordinates
}
\examples{
\dontrun{
SpatialFeaturePlot(srt, features = c("CD3D", "CD8A"))
}
}
