% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uv_create_env}
\alias{uv_create_env}
\title{Create UV virtual environment}
\usage{
uv_create_env(python_version = "3.10")
}
\arguments{
\item{python_version}{Python version to use (e.g., "3.10")}
}
\description{
Create UV virtual environment
}
