% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uv_env_exists}
\alias{uv_env_exists}
\title{Check if UV environment exists}
\usage{
uv_env_exists()
}
\value{
Logical, TRUE if .venv exists in package directory
}
\description{
Check if UV environment exists
}
