% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{uv_python}
\alias{uv_python}
\title{Get Python executable path from UV environment}
\usage{
uv_python()
}
\value{
Path to Python executable in UV environment
}
\description{
Get Python executable path from UV environment
}
\keyword{internal}
