% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uv_sync_deps}
\alias{uv_sync_deps}
\title{Sync Python dependencies using UV}
\usage{
uv_sync_deps(extras = "all")
}
\arguments{
\item{extras}{Character vector of extra dependency groups to install}
}
\description{
Sync Python dependencies using UV
}
