% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize strings}
\usage{
capitalize(x, force_tolower = FALSE)
}
\arguments{
\item{x}{Character vector to capitalize}

\item{force_tolower}{Logical. If TRUE, converts entire string to lowercase first before capitalizing.
This is useful for converting gene names from UPPERCASE to Capitalized format (e.g., "INS" -> "Ins").
Default is FALSE.}
}
\value{
Character vector with capitalized strings
}
\description{
Capitalizes the first letter of each string and optionally converts the rest to lowercase.
}
\examples{
capitalize(c("hello", "WORLD"))
# [1] "Hello" "WORLD"

capitalize(c("INS", "GCG", "NEUROG3"), force_tolower = TRUE)
# [1] "Ins" "Gcg" "Neurog3"
}
