% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seurat_data}
\alias{get_seurat_data}
\title{Get data from Seurat object in a version-agnostic way}
\usage{
get_seurat_data(srt, layer = "data", assay = NULL)
}
\arguments{
\item{srt}{A Seurat object}

\item{layer}{The layer to extract (e.g., "counts", "data", "scale.data")}

\item{assay}{Name of the assay. If NULL, uses the default assay.}
}
\value{
A matrix of the requested data
}
\description{
This function extracts data from a Seurat object, handling both V4 and V5 formats.
}
