% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invoke}
\alias{invoke}
\title{Invoke a function with a list of arguments}
\usage{
invoke(.fn, .args = list(), ...)
}
\arguments{
\item{.fn}{Function to call, can be a function object or function name as string}

\item{.args}{List of arguments to pass to the function}

\item{...}{Additional arguments passed to the function}
}
\value{
Result of the function call
}
\description{
This is a utility function to call a function with arguments provided as a list.
Similar to do.call but with more flexibility for parameter handling.
}
