% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invoke}
\alias{invoke}
\title{Invoke a function with arguments}
\usage{
invoke(.fn, .args = list(), ...)
}
\arguments{
\item{.fn}{Function or character name of function to call}

\item{.args}{List of arguments to pass to function}

\item{...}{Additional arguments (merged with .args)}
}
\description{
Wrapper around do.call that handles character function names
}
\keyword{internal}
