% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PrepareEnv}
\alias{PrepareEnv}
\title{Install the SCP python environment}
\usage{
PrepareEnv(
  method = "auto",
  pip = FALSE,
  user = FALSE,
  force = FALSE,
  update = "all",
  python_version = "3.10"
)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically find a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows.}

\item{pip}{Whether to use pip for installing packages. This is only relevant when `conda = "auto".}

\item{user}{Whether to install packages into a user site library, instead of the default system site library. Note that this argument is only relevant when using the Conda package manager. This is only relevant when `pip = TRUE`.}

\item{force}{Force reinstall the SCP python environment.}

\item{update}{Whether to update packages that already exist within the environment. Note that this parameter is only supported on Conda package manager. The available options are "default", "all", and "selected". "default" only updates the selected packages to the latest versions specified. "all" Update all packages in the environment. An alternative is "selected" to update selected packages to the latest versions while leaving all other packages. within the environment alone.}

\item{python_version}{Python version to use for the SCP environment. Default is "3.10".}
}
\description{
Install all python packages in the SCP environment automatically. The environment will be created
using a Conda environment.
}
