% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TestPythonCompatibility}
\alias{TestPythonCompatibility}
\title{Test Python Compatibility}
\usage{
TestPythonCompatibility(
  versions = c("3.10", "3.11", "3.12"),
  test_packages = TRUE
)
}
\arguments{
\item{versions}{Character vector of Python versions to test. Default is c("3.10", "3.11", "3.12").}

\item{test_packages}{Logical. Whether to test package installation. Default is TRUE.}
}
\description{
Test compatibility with different Python versions for SCP environment setup.
This function tests the specified Python versions on different platforms.
}
